+++ src/array.ts
@@ -2,11 +2,17 @@ import type { Comparator } from './types';
 import { comparisons as comps } from './utils';
 import { isValidInteger } from './validnumber';
 
-function isArray<T = any>(a: any, comp?: Comparator|null, len?: number|null): a is Array<T> {
-  return a?.constructor === Array ? (comp && isValidInteger(len, 0, 999999999)) ? comps.hasOwnProperty(comp) ? comps[comp](a.length, len) : false : true : false;
+function isArray<T = unknown>(a: unknown, comp?: Comparator|null, len?: number|null): a is Array<T> {
+  return a?.constructor === Array 
+    ? (comp && isValidInteger(len, 0, 999999999)) 
+      ? Object.prototype.hasOwnProperty.call(comps, comp) 
+        ? comps[comp]((a as Array<T>).length, len) 
+        : false 
+      : true 
+    : false;
 }
 
-function isArrayOfLength<T = any>(a: any, 
+function isArrayOfLength<T = unknown>(a: unknown, 
                                min = -999999999, 
                                max = 999999999): a is Array<T> {
   if (isArray(a, null, null)) {
@@ -16,7 +22,7 @@ function isArrayOfLength<T = any>(a: any,
   return false;
 }
 
-function isIn(val: any, arr: any[]): boolean {
+function isIn(val: unknown, arr: unknown[]): boolean {
   return isArray(arr, '>', 0) ? arr.includes(val) : false;
 }