+++ src/ch.d.ts
@@ -10,59 +10,63 @@ type PasswordOptions = {
 };
 export type { Comparator, PasswordOptions };
 
-declare function isObject<T = any>(o: any, empty?: boolean): o is object & T;
-declare function isProperty(val: any, obj: { [key: string]: any }): boolean;
+declare function isObject<T = unknown>(o: unknown, empty?: boolean): o is object & T;
+declare function isProperty<K extends PropertyKey>(v: K, obj: {
+  [key: PropertyKey]: unknown;
+}, own?: boolean, enumerable?: boolean): obj is Record<K, {
+  [key: PropertyKey]: unknown;
+}>;
 
-declare function isNil(n: any): n is null | undefined;
+declare function isNil(n: unknown): n is null | undefined;
 
-declare function isBoolean(b: any): b is boolean;
-declare function isNumber(n: any, type?: boolean): n is number;
-declare function isSymbol(s: any): s is symbol;
+declare function isBoolean(b: unknown): b is boolean;
+declare function isNumber(n: unknown, type?: boolean): n is number;
+declare function isSymbol(s: unknown): s is symbol;
 
-declare function isFunction(func: any): func is Function;
+declare function isFunction(func: unknown): func is (...args: unknown[]) => unknown;
 
-declare function isArray<T = any>(a: any, comp?: Comparator | null, len?: number | null): a is Array<T>;
-declare function isArrayOfLength<T = any>(a: any, min?: number, max?: number): a is Array<T>;
-declare function isIn(val: any, arr: any[]): boolean;
+declare function isArray<T = unknown>(a: unknown, comp?: Comparator | null, len?: number | null): a is Array<T>;
+declare function isArrayOfLength<T = unknown>(a: unknown, min?: number, max?: number): a is Array<T>;
+declare function isIn(val: unknown, arr: unknown[]): boolean;
 
-declare function isAscii(c: any, ext?: boolean): c is number;
-declare function isInteger(n: any, type?: boolean): n is number;
-declare function isFloat(n: any, type?: boolean): n is number;
-declare function isEven(n: any, type?: boolean): n is number;
-declare function isOdd(n: any, type?: boolean): n is number;
-declare function isOrigin(n: any, type?: boolean): n is number;
-declare function isPositive(n: any, type?: boolean): n is number;
-declare function isNegative(n: any, type?: boolean): n is number;
-declare function isPowerOfTwo(n: any, type?: boolean): n is number;
+declare function isAscii(c: unknown, ext?: boolean): c is number;
+declare function isInteger(n: unknown, type?: boolean): n is number;
+declare function isFloat(n: unknown, type?: boolean): n is number;
+declare function isEven(n: unknown, type?: boolean): n is number;
+declare function isOdd(n: unknown, type?: boolean): n is number;
+declare function isOrigin(n: unknown, type?: boolean): n is number;
+declare function isPositive(n: unknown, type?: boolean): n is number;
+declare function isNegative(n: unknown, type?: boolean): n is number;
+declare function isPowerOfTwo(n: unknown, type?: boolean): n is number;
 
-declare function isValidNumber(n: any, min?: number, max?: number, type?: boolean): n is number;
-declare function isValidInteger(n: any, min?: number, max?: number, type?: boolean): n is number;
-declare function isValidFloat(n: any, min?: number, max?: number, type?: boolean): n is number;
+declare function isValidNumber(n: unknown, min?: number, max?: number, type?: boolean): n is number;
+declare function isValidInteger(n: unknown, min?: number, max?: number, type?: boolean): n is number;
+declare function isValidFloat(n: unknown, min?: number, max?: number, type?: boolean): n is number;
 
-declare function isHtmlElement(h: any): h is HTMLElement;
-declare function isHtmlEventAttribute(h: any): h is string;
-declare function isNode(n: any): n is Node;
+declare function isHtmlElement(h: unknown): h is HTMLElement;
+declare function isHtmlEventAttribute(h: unknown): h is string;
+declare function isNode(n: unknown): n is Node;
 
-declare function isString(s: any, required?: boolean): s is string;
-declare function isStringOfLength(s: any, min?: number, max?: number): s is string;
-declare function isJson(s: any): s is JSON;
-declare function isRegex(r: any, type?: boolean): r is RegExp;
-declare function isEmail(e: any): e is string;
-declare function isIpAddress(i: any): i is string;
-declare function isBase64(s: any, urlEncoded?: boolean): boolean 
-declare function isJWT(t: any): t is string;
-declare function isSlug(s: any): s is string;
-declare function isHexadecimal(s: any): s is string;
-declare function containsUpperCase(s: any): s is string;
-declare function containsLowerCase(s: any): s is string;
-declare function containsSpecialCharacter(s: any): s is string;
-declare function containsNumber(s: any, min?: number | null, max?: number | null): s is string;
-declare function isValidPassword(s: any, options?: PasswordOptions): s is string;
+declare function isString(s: unknown, required?: boolean): s is string;
+declare function isStringOfLength(s: unknown, min?: number, max?: number): s is string;
+declare function isJson(s: unknown): s is JSON;
+declare function isRegex(r: unknown, type?: boolean): r is RegExp;
+declare function isEmail(e: unknown): e is string;
+declare function isIpAddress(i: unknown): i is string;
+declare function isBase64(s: unknown, urlEncoded?: boolean): boolean 
+declare function isJWT(t: unknown): t is string;
+declare function isSlug(s: unknown): s is string;
+declare function isHexadecimal(s: unknown): s is string;
+declare function containsUpperCase(s: unknown): s is string;
+declare function containsLowerCase(s: unknown): s is string;
+declare function containsSpecialCharacter(s: unknown): s is string;
+declare function containsNumber(s: unknown, min?: number | null, max?: number | null): s is string;
+declare function isValidPassword(s: unknown, options?: PasswordOptions): s is string;
 
-declare function isDate(d: any): d is Date;
-declare function isValidDate(d: any, min?: Date, max?: Date): d is Date;
-declare function isTimestamp(t: any, type?: boolean): t is number;
-declare function isValidTimestamp(t: any, min?: number, max?: number, type?: boolean): t is number;
+declare function isDate(d: unknown): d is Date;
+declare function isValidDate(d: unknown, min?: Date, max?: Date): d is Date;
+declare function isTimestamp(t: unknown, type?: boolean): t is number;
+declare function isValidTimestamp(t: unknown, min?: number, max?: number, type?: boolean): t is number;
 
 declare function ucfirst(s: string, everyWords?: boolean): string | false;
 declare function normalizeNickname(nickname: string, firstName: string, lastName: string): string | false;