+++ src/function.ts
@@ -1,6 +1,6 @@
 import { getTag } from './internal';
 
-function isFunction(func: any): func is Function {
+function isFunction(func: unknown): func is (...args: unknown[]) => unknown {
   return Boolean(func && getTag(func) === "[object Function]");
 }