+++ src/object.ts
@@ -2,16 +2,18 @@ import { isArray } from './array';
 import { isString } from './string';
 import { isNumber, isSymbol } from './primitive';
 
-function isObject<T = any>(o: any, empty = false): o is object & T {
+function isObject<T = unknown>(o: unknown, empty = false): o is object & T {
   return o !== null && typeof o === "object" && !isArray(o) && (empty ? !!Object.keys(o).length : true);
 }
 
 // own: boolean - whether to check inherited properties only
 // enumerable: boolean - whether to check enumerable properties only
-function isProperty<K extends PropertyKey>(v: K, obj: { [key: PropertyKey]: any }, own = true, enumerable = true): obj is Record<K, { [key: PropertyKey]: any }> {
+function isProperty<K extends PropertyKey>(
+  v: K, obj: { [key: PropertyKey]: unknown }, 
+  own = true, 
+  enumerable = true): obj is Record<K, { [key: PropertyKey]: unknown }> {
   
   if ((!isString(v, true) && !isNumber(v, true) && !isSymbol(v)) || !isObject(obj))
-    
     return false;
   
   if (!(v in obj)) // check broadly for property