+++ tstests/src/app.ts
@@ -5,19 +5,19 @@ declare function isProperty2<K extends PropertyKey>(
   val: K,
   obj: unknown,
 ): obj is Record<K, unknown>;
-declare function isArray2(a: unknown): a is readonly unknown[];
+// declare function isArray2(a: unknown): a is readonly unknown[];
  
-function essaiArray(a?: readonly string[] | { a: true } | null) {
-  if (isArray(a)) {
-    // KO: type is any[]
-    const b = a;
-  }
+// function essaiArray(a?: readonly string[] | { a: true } | null) {
+//   if (isArray(a)) {
+//     // KO: type is any[]
+//     const b = a;
+//   }
  
-  if (isArray2(a)) {
-    // OK: type is readonly string[]
-    const b = a;
-  }
-}
+//   if (isArray2(a)) {
+//     // OK: type is readonly string[]
+//     const b = a;
+//   }
+// }
  
 function essaiProperty(a: { a: boolean } | { b: boolean }) {
   if (isProperty("a", a)) {
@@ -41,15 +41,15 @@ function essaiObject(a?: boolean | { readonly a: boolean } | null) {
   }
 }
  
-function essaiObject2(a?: { readonly a: boolean } | null) {
-  if (isObject(a)) {
-    // KO: b is { readonly a: boolean;} | null | undefined
-    const b = a;
-  }
-  if (isObject2(a)) {
-    // OK: b is { readonly a: boolean;}
-    const b = a;
-  }
-}
+// function essaiObject2(a?: { readonly a: boolean } | null) {
+//   if (isObject(a)) {
+//     // KO: b is { readonly a: boolean;} | null | undefined
+//     const b = a;
+//   }
+//   if (isObject2(a)) {
+//     // OK: b is { readonly a: boolean;}
+//     const b = a;
+//   }
+// }